Refactor AlterShardingBindingTableRuleStatementUpdaterTest
